Refactor: Use create-checkout-session API endpoint
