+++ fix-404/app/booking/page.tsx.backup
@@ -187,14 +187,12 @@ export default function BookingPage() {
       const chosenService =
         servicePricing[selectedService as keyof typeof servicePricing]
 
-      const response = await fetch("/api/create-payment-intent", {
+      const response = await fetch("/api/create-checkout-session", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
         },
         body: JSON.stringify({
-          amount: calculateTotal() * 100,
-          currency: "cad",
           service: chosenService,
           addons: selectedAddons.map((id) => addons.find((a) => a.id === id)).filter(Boolean),
           customerInfo: formData,
